# Improve performance by removing sync-over-async by generating sync methods using Zomp.SyncMethodGenerator