--- a/src/FluentValidation/AbstractValidator.cs
+++ b/src/FluentValidation/AbstractValidator.cs
@@ -137,7 +137,7 @@ public abstract partial class AbstractValidator<T> : IValidator<T>, IEnumerable<
 		return await ValidateInternalAsync(context, cancellation);
 	}
 
-	[Zomp.SyncMethodGenerator.CreateSyncVersion]
+	[Zomp.SyncMethodGenerator.CreateSyncVersion(OmitNullableDirective = true)]
 	private async ValueTask<ValidationResult> ValidateInternalAsync(ValidationContext<T> context, CancellationToken cancellation) {
 		var result = new ValidationResult(context.Failures);
 		bool shouldContinue = PreValidate(context, result);