--- a/src/FluentValidation/Internal/PropertyRule.cs
+++ b/src/FluentValidation/Internal/PropertyRule.cs
@@ -28,7 +28,7 @@ using System.Threading.Tasks;
 /// <summary>
 /// Defines a rule associated with a property.
 /// </summary>
-internal class PropertyRule<T, TProperty> : RuleBase<T, TProperty, TProperty>, IValidationRuleInternal<T, TProperty> {
+internal partial class PropertyRule<T, TProperty> : RuleBase<T, TProperty, TProperty>, IValidationRuleInternal<T, TProperty> {
 
 	public PropertyRule(MemberInfo member, Func<T, TProperty> propertyFunc, LambdaExpression expression, Func<CascadeMode> cascadeModeThunk, Type typeToValidate)
 		: base(member, propertyFunc, expression, cascadeModeThunk, typeToValidate) {
@@ -47,15 +47,9 @@ internal class PropertyRule<T, TProperty> : RuleBase<T, TProperty, TProperty>, I
 	/// Performs validation using a validation context and adds collected validation failures to the Context.
 	/// </summary>
 	/// <param name="context">Validation Context</param>
-	/// <param name="useAsync">
-	/// Whether asynchronous components are allowed to execute.
-	/// This will be set to True when ValidateAsync is called on the root validator.
-	/// This will be set to False when Validate is called on the root validator.
-	/// When set to True, asynchronous components and asynchronous conditions will be executed.
-	/// When set to False, an exception will be thrown if a component can only be executed asynchronously or if a component has an async condition associated with it.
-	/// </param>
 	/// <param name="cancellation"></param>
-	public virtual async ValueTask ValidateAsync(ValidationContext<T> context, bool useAsync, CancellationToken cancellation) {
+	[Zomp.SyncMethodGenerator.CreateSyncVersion]
+	public virtual async ValueTask ValidateAsync(ValidationContext<T> context, CancellationToken cancellation) {
 		string displayName = GetDisplayName(context);
 
 		if (PropertyName == null && displayName == null) {
@@ -79,14 +73,12 @@ internal class PropertyRule<T, TProperty> : RuleBase<T, TProperty, TProperty>, I
 		}
 
 		if (AsyncCondition != null) {
-			if (useAsync) {
 				if (!await AsyncCondition(context, cancellation)) {
 					return;
 				}
-			}
-			else {
+#if SYNC_ONLY
 				throw new AsyncValidatorInvokedSynchronouslyException();
-			}
+#endif
 		}
 
 		bool first = true;
@@ -107,14 +99,12 @@ internal class PropertyRule<T, TProperty> : RuleBase<T, TProperty, TProperty>, I
 			}
 
 			if (component.HasAsyncCondition) {
-				if (useAsync) {
-					if (!await component.InvokeAsyncCondition(context, cancellation)) {
-						continue;
-					}
+				if (!await component.InvokeAsyncCondition(context, cancellation)) {
+					continue;
 				}
-				else {
-					throw new AsyncValidatorInvokedSynchronouslyException();
-				}
+#if SYNC_ONLY
+				throw new AsyncValidatorInvokedSynchronouslyException();
+#endif
 			}
 
 			if (first) {
@@ -127,7 +117,7 @@ internal class PropertyRule<T, TProperty> : RuleBase<T, TProperty, TProperty>, I
 				}
 			}
 
-			bool valid = await component.ValidateAsync(context, propValue, useAsync, cancellation);
+			bool valid = await component.ValidateAsync(context, propValue, cancellation);
 
 			if (!valid) {
 				PrepareMessageFormatterForValidationError(context, propValue);
@@ -145,7 +135,7 @@ internal class PropertyRule<T, TProperty> : RuleBase<T, TProperty, TProperty>, I
 		if (context.Failures.Count <= totalFailures && DependentRules != null) {
 			foreach (var dependentRule in DependentRules) {
 				cancellation.ThrowIfCancellationRequested();
-				await dependentRule.ValidateAsync(context, useAsync, cancellation);
+				await dependentRule.ValidateAsync(context, cancellation);
 			}
 		}
 	}