The main advantage of this approach is performance.

The performance is increased by 26%, 21%, 22%, 28%, 30% and 7% for the following benchmarks in ValidationBenchmark.

### useAsync flag

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.22621
AMD Ryzen 9 5900HS with Radeon Graphics, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=7.0.400-preview.23330.10
  [Host]     : .NET Core 6.0.20 (CoreCLR 6.0.2023.32017, CoreFX 6.0.2023.32017), X64 RyuJIT
  DefaultJob : .NET Core 6.0.20 (CoreCLR 6.0.2023.32017, CoreFX 6.0.2023.32017), X64 RyuJIT


|   Method |    DataSet |      Mean |     Error |     StdDev |      Gen 0 |     Gen 1 | Gen 2 | Allocated |
|--------- |----------- |----------:|----------:|-----------:|-----------:|----------:|------:|----------:|
| FailFast | HalfErrors | 452.61 ms | 19.804 ms |  57.456 ms | 31000.0000 |         - |     - | 249.08 MB |
| Validate | HalfErrors | 613.89 ms | 23.585 ms |  68.425 ms | 45000.0000 |         - |     - | 362.54 MB |
| FailFast | ManyErrors |  23.21 ms |  0.981 ms |   2.766 ms |  2562.5000 |         - |     - |   20.5 MB |
| Validate | ManyErrors | 996.51 ms | 41.464 ms | 120.952 ms | 55000.0000 | 3000.0000 |     - | 444.06 MB |
| FailFast |   NoErrors | 865.26 ms | 29.054 ms |  83.360 ms | 44000.0000 |         - |     - | 355.72 MB |
| Validate |   NoErrors | 678.55 ms | 42.455 ms | 119.745 ms | 44000.0000 |         - |     - | 355.72 MB |


### SyncMethodGenerator

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.22621
AMD Ryzen 9 5900HS with Radeon Graphics, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=7.0.400-preview.23330.10
  [Host]     : .NET Core 6.0.20 (CoreCLR 6.0.2023.32017, CoreFX 6.0.2023.32017), X64 RyuJIT
  DefaultJob : .NET Core 6.0.20 (CoreCLR 6.0.2023.32017, CoreFX 6.0.2023.32017), X64 RyuJIT


|   Method |    DataSet |      Mean |     Error |     StdDev |    Median |      Gen 0 |     Gen 1 | Gen 2 | Allocated |
|--------- |----------- |----------:|----------:|-----------:|----------:|-----------:|----------:|------:|----------:|
| FailFast | HalfErrors | 333.36 ms | 19.547 ms |  57.329 ms | 322.04 ms | 31000.0000 |         - |     - | 249.08 MB |
| Validate | HalfErrors | 484.47 ms | 23.499 ms |  68.547 ms | 475.97 ms | 45000.0000 |         - |     - | 362.54 MB |
| FailFast | ManyErrors |  18.55 ms |  0.754 ms |   2.150 ms |  17.90 ms |  2562.5000 |         - |     - |   20.5 MB |
| Validate | ManyErrors | 713.78 ms | 36.968 ms | 108.420 ms | 693.34 ms | 55000.0000 | 3000.0000 |     - | 444.06 MB |
| FailFast |   NoErrors | 603.87 ms | 25.970 ms |  75.756 ms | 602.52 ms | 44000.0000 |         - |     - | 355.72 MB |
| Validate |   NoErrors | 632.17 ms | 39.229 ms | 115.053 ms | 626.33 ms | 44000.0000 |         - |     - | 355.72 MB |

Other than that, there is less lines of code and no "acceptable" sync-over-async, which still generates a state machine. It was also throwing me off seeing "Async" (eg ValidateAsync and ValidateInternalAsync) in the stack trace when I wasn't doing anything asynchronously.

The only downside is that the DLL grew from 458 KB to 466  KB, but that's less than 2%.