@JeremySkinner 2023-08-08
Thanks for the suggestion - this is an interesting idea. My initial reaction was to say that this isn't something I'm keen on, but  thinking about it further I'm open to have a discussion about it but I have a few concerns:

- Makes it much less obvious what's happening, I prefer the explicitness of the current approach.
- I want to avoid forcing consumers to take on additional dependency. Does this work with `PrivateAssets=all` in the package reference? 
- Build fails currently with the error `Zomp.SyncMethodGenerator.dll' references version '4.6.0.0' of the compiler, which is newer than the currently running version '4.4.0.0'.`. FluentValidation compilation is done with the .NET 7 SDK, targeting  following TFMs: `netstandard2.0;netstandard2.1;net5.0;net6.0;net7.0` (although netstandard2.0 and net5.0 will be removed in the next major release).