@virzak 2023-08-08
In reverse order: library users won't know the difference other than performance improvement and perhaps stack trace. Looking at the new package using NuGet Package Explorer is this:

![image](https://github.com/FluentValidation/FluentValidation/assets/6209775/4a33012c-f661-4621-be17-23c127059bea)

It seems that no source generator needs to add `PrivateAssets=all`


About the code being less obvious. Since library users won't know the difference, the question is how less obvious is it for contributors?

You would be correct to say that any source generator is less obvious, because sometimes you can't see a method / property unless you F12 from Visual Studio. Previous contributors might wonder what is `CreateSyncVersion` attribute.

On the other hand, this sync-over-async approach also had to have a paragraph with explanation of why it is acceptable and any new contributor who knows that Async shouldn't be used in synchronized methods will have a learning curve. Which is the steeper learning curve? Also the end users who see 'Async' in the stack trace might be confused. Btw, that's the exact reason for me submitting this PR.

IMO, while code clarity and obviousness can be argued either way, the boost in performance tips the scale for this change.

I'd be curious to know what was the most confusing part for you in this approach, so perhaps it could be remedied.

For example one idea I had (which isn't yet implemented):

Instead of:
```cs
#if SYNC_ONLY
    throw new AsyncValidatorInvokedSynchronouslyException();
#endif
```

The syntax could be rewritten as

```cs
[ThrowWhenInvokedSynchronously<AsyncValidatorInvokedSynchronouslyException>]
internal Func<ValidationContext<T>, CancellationToken, Task<bool>> AsyncCondition => _asyncCondition;
```