@JeremySkinner 2023-08-08
hi @virzak thanks for getting back to me

> Looking at the new package using NuGet Package Explorer is this:

What does the list of dependent packages look like? 

> On the other hand, this sync-over-async approach also had to have a paragraph with explanation of why it is acceptable and any new contributor who knows that Async shouldn't be used in synchronized methods will have a learning curve.

Yep all good points, it's a tradeoff, although personally I prefer having the explicitness of being able to see the method with an explanatory comment, and think this is more contributor friendly. 

> Also the end users who see 'Async' in the stack trace might be confused. Btw, that's the exact reason for me submitting this PR.

That could be partially solved by dropping the async suffix from the method name if its causing confusion. I'm not precious about using the async suffix for internal methods if it clears up ambiguity. 

> IMO, while code clarity and obviousness can be argued either way, the boost in performance tips the scale for this change.

At the end of the day I think it'll be a case of whether the performance boost balances out my discomfort or not. I used to maintain 2 explicit code paths (sync & async), although this technically performed better than the current approach, it was a massive pain to maintain and the minor performance degredation was balanced out by other performance savings made elsewhere. I'll be the one working on the codebase and maintaining this approach, so I need to be completely comfortable with it, so it needs to be something I'm comfortable with.

> I'd be curious to know what was the most confusing part for you in this approach, so perhaps it could be remedied.

Generally I don't like not being able to see the method implementation, although that's a broader issue with source generators as a whole, and why I don't generally like to use them as I want to actually see the code that I'm compiling. I'm also not fond of the conditional `SYNC_ONLY` symbol; if the method is now purely async then I wouldn't expect to see any sync code paths in there at all. The attribute as an alternative approach is interesting, although again the more rewriting it's doing behind the scenes the less comfortable I am with it.

Also how battle tested is this in production? It looks pretty new, and for a library like FluentValidation with such a large userbase over such a long period of time I try to prioritise stability and am very cautious about taking a dependency on anything that's not been extensively battle tested.